Title: 괄호 쌍 만들어주기, Time: 133ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240123/괄호 쌍 만들어주기/pair-parentheses.java
@@ -15,7 +15,7 @@ public class Main {
             }
         }
 
-        int output = 0;
+        long output = 0;
         for(int i = 1 ; i < stringArr.length ; i++){
             if(stringArr[i] == '(' && stringArr[i] == stringArr[i - 1]){
                 output += rightArr[i];
